+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Adds `L10n::is()` method for localized route matching.
 - Adds `L10n` facade.
+- Adds domain translations.
 
 ## v0.3.0